==> notebooks/00_exploration.ipynb <==
import kagglehub

# Download latest version
path = kagglehub.dataset_download("hassankhaled21/egyptian-real-estate-listings")

print("Path to dataset files:", path)

import pandas as pd
import numpy as np

df = pd.read_csv(path + "/egypt_real_estate_listings.csv")
df.head()

df.info()

down_payment_list = df["down_payment"].isna().sum()
print(down_payment_list)
