==> notebooks/03_feature_engineering.ipynb <==
DATASET_PATH = '../dataset/egypt_real_estate_listings_cleaned.csv'

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import plotly.express as px
from sklearn.preprocessing import StandardScaler , MultiLabelBinarizer 
from sklearn.manifold import TSNE
from statsmodels.api import OLS 
import statsmodels.api as sm
from statsmodels.api import add_constant
from scipy import stats
from sklearn.model_selection import train_test_split
import joblib

df = pd.read_csv(DATASET_PATH)

df.head()

# === Numerical Features Hypothesis Test (Pearson Correlation) ===
num_features = df.select_dtypes(include=[np.number]).columns.drop('price').tolist()
print("📊 Numerical Features Hypothesis Test (Correlation with Price)")
for feature in num_features:
    corr, p_val = stats.pearsonr(df[feature], df['price'])
    print(f"{feature}: correlation = {corr:.3f}, p-value = {p_val:.4f}")
    if p_val < 0.05:
        print(" → Reject H₀: Significant relationship with price ✅\n")
    else:
        print(" → Fail to reject H₀: No significant relationship with price ❌\n")


# === Categorical Features Hypothesis Test (ANOVA / t-test) ===
cat_features = df.select_dtypes(include=['object']).columns.tolist()
print("🧩 Categorical Features Hypothesis Test (Group Differences in Price)")
for feature in cat_features:
    groups = [df.loc[df[feature] == cat, 'price'] for cat in df[feature].unique()]
    
    # Skip if category has only one unique value
    if len(groups) < 2:
        continue

    # Choose test type
    if len(groups) == 2:
        stat, p_val = stats.ttest_ind(groups[0], groups[1], equal_var=False)
        test_type = "t-test"
    else:
        stat, p_val = stats.f_oneway(*groups)
        test_type = "ANOVA"
    
    print(f"{feature}: {test_type} p-value = {p_val:.4f}")
    if p_val < 0.05:
        print(" → Reject H₀: Mean price differs significantly between categories ✅\n")
    else:
        print(" → Fail to reject H₀: No significant mean difference ❌\n")

new_df = df.drop(columns=['payment_method'])

significant_features = new_df.columns.tolist()

significant_features

cat_features = new_df.select_dtypes(include=['object']).columns.tolist()

cat_features

df_encoded = pd.get_dummies(df[significant_features], drop_first=True)

df_encoded

num_cols = df_encoded.select_dtypes(include=[np.number]).columns.tolist()

num_cols

scaler = StandardScaler()
df_encoded[num_cols] = scaler.fit_transform(df_encoded[num_cols])

X = df_encoded
y = df['price']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42,shuffle=True)


X_train.shape, X_test.shape, y_train.shape, y_test.shape 

X_train

# Save preprocessors and model
joblib.dump(scaler, "../models/scaler.pkl")
joblib.dump(df_encoded.columns, "../models/model_columns.pkl")
